Add new struct for project
